+++ CHANGELOG.md
@@ -5,11 +5,13 @@
 ## 1.4.4
 - Feat: Delete reading records of a book
 - Feat: Add webdav sync direction dialog
+- Feat: Add font delete
 - Fix: Webdav sync aborted dialog content
 - Fix: if webdav is empty, sync will upload
 
 - 新增：删除一本书的阅读记录
 - 新增：添加 WebDAV 同步方向对话框
+- 新增：添加字体删除
 - 修复：WebDAV 同步中止对话框内容
 - 修复：如果 WebDAV 为空，则同步时默认上传